# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Kredit Plus backend services.",
        "title": "BE Kredit Plus",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/loan": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Retrieves available loans.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Get loan",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Per Page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful loan list response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.LoanResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized error response",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Creates a new loan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Create loan",
                "parameters": [
                    {
                        "description": "Create Loan Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful loan creation response",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error response",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error response",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error response",
                        "schema": {
                            "$ref": "#/definitions/model.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Message!"
                }
            }
        },
        "model.CreateLoanRequest": {
            "type": "object",
            "required": [
                "assets_name",
                "otr",
                "tenor_months"
            ],
            "properties": {
                "assets_name": {
                    "type": "string",
                    "example": "Credit Motorcycle"
                },
                "otr": {
                    "type": "number",
                    "minimum": 0,
                    "example": 15000000
                },
                "tenor_months": {
                    "type": "integer",
                    "maximum": 32,
                    "minimum": 1,
                    "example": 18
                }
            }
        },
...
        "model.LoanResponse": {
            "type": "object",
            "properties": {
                "admin_fee": {
                    "type": "number"
                },
                "assets_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "installment_amount": {
                    "type": "number"
                },
                "otr": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "tenor_months": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_paid": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}